code/pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "dlt",
    "dlt[postgres]",
    "dlt[workspace]",
    "python-dotenv>=1.2.1",
    "openpyxl>=3.0.0",
    "alive-progress>=3.3.0",
    "marimo",
    "numpy",
    #"numpy==2.3.5",
    "dbt-core>=1.11.2",
    "dbt-postgres>=1.1.10",
    "psycopg[binary]>=3.2.10",
    "typesense>=1.1.1",
    "dbt-osmosis",
    "shandy-sqlfmt[jinjafmt]",
]

[dependency-groups]
dev = [
    "ty>=0.0.7",
]
